Move Trends Links under courses/
