+++ new-dti-website/components/courses/data/config.json
@@ -1,0 +1,4 @@
+{
+  "trendsApplicationLink": "https://bit.ly/TrendsFA24",
+  "trendsWebsiteLink": "https://webdev.cornelldti.org"
+}
