+++ new-dti-website/config.json
@@ -4,8 +4,6 @@
   "applicationLink": "https://forms.gle/F9tFrW88a2Xoa9D49",
   "coffeeChatLink": "https://docs.google.com/spreadsheets/d/1jDrkQ61h41w0kw6gITUJJiJ-yQtE_WfBZbHvVow17ik",
   "coffeeChatFormLink": "https://docs.google.com/forms/d/e/1FAIpQLSdAX0T_H0h9qpXSo6aeM-azFbAgIfOCOlXByHpf2HkHi8XTnw/viewform",
-  "trendsApplicationLink": "https://bit.ly/TrendsFA24",
-  "trendsWebsiteLink": "https://webdev.cornelldti.org",
   "events": [
     {
       "title": "Applications Open!",
